+++ CHANGELOG.md
@@ -1,6 +1,36 @@
 CHANGELOG
 ===============
 
+*V 0.2.1*
++ Input item with `withInputItem(...)`
+
+*V 0.2.0*
++ Spinner item with `withSpinnerItem(...)`
++ Restart App item
++ Restart Activity item
+
+*V 0.1.8*
+- Unused button "Shake to"
+
+*V 0.1.7*
+= min SDK version 15.
+
+*V 0.1.6 BROKEN BUILD* 
+= Add scalpel layout with `.withScalpelLayout` instead of using interface.
+
+*V 0.1.5*
++ Clean Debugger init.
++ Proguard setup. 
+
+*V 0.1.2 && 0.1.3*
++ Manifest with needed setup
+
+*V 0.1.1*
++ Proguard setup
+
+*V 0.1*
++ First release with all the apps using it.
+
 *V 0.0.1-SNAPSHOT*
 + Debug Drawer.
 + ViewServer.